COM-3840: move navigation logic in component
