+++ src/components/AttendanceSheetDataSelectionForm/styles.ts
@@ -1,8 +1,19 @@
 import { StyleSheet } from 'react-native';
-import { MARGIN } from '../../styles/metrics';
+import { MARGIN, PADDING } from '../../styles/metrics';
 import { FIRA_SANS_BOLD } from '../../styles/fonts';
+import { GREY } from '../../styles/colors';
 
 const styles = StyleSheet.create({
+  safeArea: {
+    flex: 1,
+    backgroundColor: GREY[100],
+  },
+  header: {
+    paddingTop: PADDING.LG,
+    paddingHorizontal: PADDING.LG,
+    alignItems: 'center',
+    flexDirection: 'row',
+  },
   container: {
     flexGrow: 1,
     marginHorizontal: MARGIN.LG,
