+++ src/components/UploadMethods/styles.ts
@@ -1,7 +1,18 @@
 import { StyleSheet } from 'react-native';
-import { MARGIN } from '../../styles/metrics';
+import { MARGIN, PADDING } from '../../styles/metrics';
+import { GREY } from '../../styles/colors';
 
 const styles = StyleSheet.create({
+  safeArea: {
+    flex: 1,
+    backgroundColor: GREY[100],
+  },
+  header: {
+    paddingTop: PADDING.LG,
+    paddingHorizontal: PADDING.LG,
+    alignItems: 'center',
+    flexDirection: 'row',
+  },
   container: {
     flex: 1,
     justifyContent: 'center',
